fixing the error for overflow

+++ C++09/ex01/RPN.hpp
@@ -6,7 +6,7 @@
 /*   By: sbartoul <sbartoul@student.42abudhabi.ae>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/25 18:05:36 by sbartoul          #+#    #+#             */
-/*   Updated: 2025/01/26 01:11:22 by sbartoul         ###   ########.fr       */
+/*   Updated: 2025/02/05 14:28:07 by sbartoul         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,20 +19,21 @@
 #include <stdexcept>
 #include <sstream>
 #include <cstdlib>
+#include <climits>
 
 class RPN {
 private:
-	std::stack<int> _stack;
+	std::stack<long long> _stack;
 	std::string _input;
 	bool isOperator(const std::string& token) const;
-	int performOperation(int a, int b, const std::string& op) const;
+	long long performOperation(long long a, long long b, const std::string& op) const;
 public:
 	RPN();
 	RPN(const std::string& input);
 	RPN(const RPN& old);
 	RPN& operator=(const RPN& rhs);
 	~RPN();
-	int evaluate();
+	long long evaluate();
 };
 
 #endif
